Move pluginsArray into helpers dir
