+++ packages/core/src/js/helpers/pluginsArray.js
